GameManager fix
