--- Akuru/Assets/SSM/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e0fb546c54615a14d96fee11e10090bf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
