+++ Akuru/Assets/SSM/Scripts/Customer.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
 
 public class Customer : MonoBehaviour
 {
