+++ Akuru/Assets/SSM/Scripts/CustomerList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0fb817c916b5989499d29107e3ad657a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
