+++ Akuru/Assets/SSM/Scripts/StandPurchase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 721850308a16f0844b47f1fafd084015
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
