+++ Akuru/Assets/SSM/Scripts/Tanghulu.cs
@@ -4,8 +4,7 @@ using UnityEngine;
 
 public class Tanghulu : MonoBehaviour
 {
-    // 탕후루 가격 계산 시간 맛은 인스펙터에서 각각 써놓기
+    // 탕후루 가격 계산 시간은 인스펙터에서 각각 써놓기
     public float price;
     public float payDelay;
-    public string flavor;
 }
